programs/oracle/programs/signal-ingestion/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "signal-ingestion"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "oracle-shared",
    "redis[hiredis]>=5.0",
    "httpx>=0.27",
    "py-clob-client>=0.18",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "praw>=7.7",
    "web3>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["signal_ingestion"]
